Convert gpx format to a derived class of Format.
