+++ GPSBabel.pro
@@ -110,6 +110,7 @@ HEADERS =  \
 	gbser.h \
 	gbser_private.h \
 	gbversion.h \
+	gpx.h \
 	grtcirc.h \
 	heightgrid.h \
 	holux.h \
